requirements.txt
Flask==3.1.0
injector==0.22.0
openai==1.70.0
python-dotenv==1.1.0
